@@ -1,4 +1,4 @@
-// swift-tools-version:5.5
+// swift-tools-version:5.7
 
 import PackageDescription
 
@@ -18,10 +18,10 @@ let package = Package(
     ],
     dependencies: [
         .package(url: "https://github.com/apple/swift-async-algorithms", branch: "main"),
-        .package(url: "https://github.com/vmanot/Diagnostics.git", .branch("master")),
-        .package(url: "https://github.com/vmanot/FoundationX.git", .branch("master")),
-        .package(url: "https://github.com/vmanot/Swallow.git", .branch("master")),
-        .package(url: "https://github.com/SwiftUIX/SwiftUIX.git", .branch("master"))
+        .package(url: "https://github.com/vmanot/Diagnostics.git", branch: "master"),
+        .package(url: "https://github.com/vmanot/FoundationX.git", branch: "master"),
+        .package(url: "https://github.com/vmanot/Swallow.git", branch: "master"),
+        .package(url: "https://github.com/SwiftUIX/SwiftUIX.git", branch: "master")
     ],
     targets: [
         .target(
